% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_scalegram}
\alias{plot_scalegram}
\title{Plots scalegram}
\usage{
plot_scalegram(df, MODE, transparancy)
}
\arguments{
\item{MODE}{the statistic used.}

\item{x}{A vector, time series or a matrix.}
}
\value{
A list with the scalegram of \code{x} for statistic \code{MODE} and the corresponding plot [ggplot object].
}
\description{
Plots scalegram
}
\examples{
scalegram(dataset, "s1")
scalegram(dataset[,1], "s2")
}
