% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale_variance}
\alias{rescale_variance}
\title{Rescales variance}
\usage{
rescale_variance(emp_scalegram_coarse, emp_scalegram_fine, scale_ratio)
}
\arguments{
\item{x}{A vector, time series or a matrix.}

\item{MODE}{the statistic used.}
}
\value{
A list with the scalegram of \code{x} for statistic \code{MODE} and the corresponding plot [ggplot object].
}
\description{
Rescales variance
}
\examples{
scalegram(dataset, "s1")
scalegram(dataset[,1], "s2")
}
