% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalegram.R
\name{scalegram}
\alias{scalegram}
\title{Estimates scalegram(s) of specific statistics and plots the result.}
\usage{
scalegram(x, stat = "s1", std = TRUE, threshold = 30, plot = TRUE)
}
\arguments{
\item{x}{A vector, time series or a matrix.}

\item{stat}{The statistic used (see Details).}

\item{std}{If TRUE (default) standardize the scalegram to unit.}

\item{threshold}{Sample size at last scale (see Details).}

\item{plot}{If TRUE (default) the scalegram is also plotted.}
}
\value{
A list with the scalegram of \code{x} for statistic \code{stat} and the corresponding plot [ggplot object].
}
\description{
Estimates scalegram(s) of specific statistics and plots the result.
}
\examples{
scalegram(dataset, "s1")
scalegram(dataset[,1], "s2")
}
