% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scalegram_main}
\alias{scalegram_main}
\title{Main scalegram function}
\usage{
scalegram_main(x, MODE, STD, threshold)
}
\arguments{
\item{x}{A vector, time series or a matrix.}

\item{MODE}{the statistic used.}
}
\value{
A list with the scalegram of \code{x} for statistic \code{MODE} and the corresponding plot [ggplot object].
}
\description{
Main scalegram function
}
\examples{
scalegram(dataset, "s1")
scalegram(dataset[,1], "s2")
}
