% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalegram.R
\name{scalegram}
\alias{scalegram}
\title{Estimate and plot scalegram}
\usage{
scalegram(x, stat = "sd", std = TRUE, threshold = 30, plot = TRUE)
}
\arguments{
\item{x}{The dataset given as vector, time series or a matrix.}

\item{stat}{The statistic which will be estimated across the scale continuum. Possible statistics are:
\itemize{
 \item{"mean" for mean,}
 \item{"sd" for standard deviation,}
 \item{"var" for variance,}
 \item{"skew" for skewness,}
 \item{"kurt" for kurtosis,}
 \item{"cv" for coefficient of variance,}
 \item{"L2" for L-scale,}
 \item{"t2" for coefficient of L-variation,}
 \item{"t3" for L-skewness,}
 \item{"t4" for L-kurtosis.}
}}

\item{std}{If TRUE (default) standardize the scalegram to unit, i.e., zero mean and unit variance at the original time scale.}

\item{threshold}{Sample size at the last scale (see Details).}

\item{plot}{If TRUE (default) the scalegram is also plotted.}
}
\value{
A list containing the scalegram of \code{x} for the given \code{stat} statistic and the corresponding plot as a \emph{ggplot object}.
}
\description{
Estimates scalegram(s) of specific statistics and plots the result.
}
\details{
Here are the details.
}
\examples{
scalegram(rnorm(1000))

## Plot scalegram in logarithmic y axis
owda_site = owda[owda$Lat == 46.25 & owda$Lon == 16.5,]$scPDSI
sgram_sd <- scalegram(owda_site, "L2")
sgram_sd$scalegram_plot + scale_y_log10("L-scale")

## Plot matrix
owda_mat = owda_mat = acast(owda, Time ~ Lon + Lat, value.var = "scPDSI")
sgram_mat_sd = scalegram(owda_mat, std = F, threshold = 50)



}
