% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalegram.R
\name{scalegram}
\alias{scalegram}
\title{Estimate and plot scalegram}
\usage{
scalegram(x, stat = "sd", std = TRUE, threshold = 30, plot = TRUE)
}
\arguments{
\item{x}{A numeric vector, time series or matrix.}

\item{stat}{The statistic which will be estimated across the cross-scale continuum. Suitable options are:
\itemize{
 \item{"sd" for standard deviation,}
 \item{"var" for variance,}
 \item{"skew" for skewness,}
 \item{"kurt" for kurtosis,}
 \item{"L2" for L-scale,}
 \item{"t2" for coefficient of L-variation,}
 \item{"t3" for L-skewness,}
 \item{"t4" for L-kurtosis.}
}}

\item{std}{logical. If TRUE (the default) the scalegram is standardized to unit, i.e., zero mean and unit variance in the original time scale.}

\item{threshold}{numeric. Sample size of the time series at the last aggregated scale (see Details).}

\item{plot}{logical. If TRUE (the default) the scalegram is plotted.}
}
\value{
A list with two elemets:
\describe{
  \item{One}{First item}
  \item{Two}{Second item}
} containing the scalegram of \code{x} for the given \code{stat} statistic
and the corresponding plot as a \emph{ggplot object}. The
}
\description{
The function \code{scalegram} computes (and by default plots) estimates of the scaling behaviour of given statistics.
}
\details{
Here are the details.
}
\examples{
scalegram(rnorm(1000))

## Plot scalegram in logarithmic y axis
owda_site = owda[owda$Lat == 46.25 & owda$Lon == 16.5,]$scPDSI
sgram_sd <- scalegram(owda_site, "L2")
sgram_sd$scalegram_plot + scale_y_log10("L-scale")

## Plot matrix
owda_mat = owda_mat = acast(owda, Time ~ Lon + Lat, value.var = "scPDSI")
sgram_mat_sd = scalegram(owda_mat, std = F, threshold = 50)



}
