% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scalegram}
\alias{scalegram}
\title{Main scalegram function}
\usage{
scalegram(x, MODE = "s1", STD = TRUE, threshold = 30, PLOT = TRUE)
}
\arguments{
\item{x}{A vector, time series or a matrix.}

\item{MODE}{the statistic used.}
}
\value{
A list with the scalegram of \code{x} for statistic \code{MODE} and the corresponding plot [ggplot object].
}
\description{
Main scalegram function
}
\examples{
scalegram(dataset, "s1")
scalegram(dataset[,1], "s2")
}
