% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{scalegram_parallel}
\alias{scalegram_parallel}
\title{Function for parallel computing used when input data are in matrix format.}
\usage{
scalegram_parallel(x, stat, std, threshold, cores_used = detectCores() - 1,
  ...)
}
\arguments{
\item{x}{A vector, time series or a matrix.}

\item{stat}{the statistic used.}
}
\value{
A list with the scalegram of \code{x} for statistic \code{stat} and the corresponding plot [ggplot object].
}
\description{
Function for parallel computing used when input data are in matrix format.
}
\examples{
scalegram(dataset, "s1")
scalegram(dataset[,1], "s2")
}
